Bug fixes for leakage calculations

+++ test/python/modules/linear_boltzmann_solvers/transport_steady/angular_io_1d.py
@@ -5,6 +5,7 @@
 
 import os
 import sys
+import numpy as np
 
 if "opensn_console" not in globals():
     from mpi4py import MPI
@@ -86,31 +87,29 @@ if __name__ == "__main__":
         "save_angular_flux": True,
     }
 
+    # Initialize and execute solvers
     phys1 = DiscreteOrdinatesProblem(**solver_dict)
-
-    # Initialize and execute solver
     ss_solver = SteadyStateSourceSolver(problem=phys1)
     ss_solver.Initialize()
     ss_solver.Execute()
-
-    leakage_left_1 = phys1.ComputeLeakage(["zmin"])["zmin"][0]
-    leakage_right_1 = phys1.ComputeLeakage(["zmax"])["zmax"][0]
+    psi1 = phys1.GetPsi()
     phys1.WriteAngularFluxes("angular_io")
 
     phys2 = DiscreteOrdinatesProblem(**solver_dict)
     ss_solver_2 = SteadyStateSourceSolver(problem=phys2)
     ss_solver_2.Initialize()
     phys2.ReadAngularFluxes("angular_io")
+    psi2 = phys2.GetPsi()
 
-    leakage_left_2 = phys2.ComputeLeakage(["zmin"])["zmin"][0]
-    leakage_right_2 = phys2.ComputeLeakage(["zmax"])["zmax"][0]
+    for i, (arr1, arr2) in enumerate(zip(psi1, psi2)):
+        if arr1.shape != arr2.shape:
+            raise ValueError(f"Different shapes at index {i}: {arr1.shape} vs {arr2.shape}")
 
-    leakage_left_diff = leakage_left_1 - leakage_left_2
-    leakage_right_diff = leakage_right_1 - leakage_right_2
+    if not np.allclose(arr1, arr2, rtol=1e-8, atol=1e-12):
+        diff = np.max(np.abs(arr1 - arr2))
+        print(f"Difference at index {i}, max diff = {diff}")
+        raise ValueError(f"psi mismatch at index {i}")
 
-    if rank == 0:
-        print(f"Leakage-Diff1={leakage_left_diff:.5e}")
-    if rank == 0:
-        print(f"Leakage-Diff2={leakage_right_diff:.5e}")
+    print(f"psi values match for rank {rank}")
 
     os.remove(f"angular_io{rank}.h5")
